Split out big chunk of hub.js into room entry manager
